requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
boto3==1.24.26
botocore==1.27.26
braintree==4.16
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-ckeditor==6.4.2
django-cors-headers==3.7.0
django-environ==0.9.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
Js2Py==0.71
Markdown==3.3.7
MarkupPy==1.14
MarkupSafe==2.1.1
mutagen==1.45.1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
pipwin==0.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pyjsparser==2.7.1
PyJWT==2.4.0
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==3.5.0
tablib==3.2.1
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.10
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
